chore: update ci and fix version (#72)
Hi @drewm102, this PR aims at solving #71 as the current repo isn't
aligned with the release version in PyPI.

The version change is performed to allow a new release 1.2.0 to be
performed. This is required to align with the fact that PyPI has version
1.0.11 available. In the following releases, you'll have to follow the
1.Y.Z version or this could backfire to your users.

The other change makes the repo use trusted publisher approach.

Another way of solving all of this would be to disable the releases
1.Y.Z from PyPI and I'd encourage to do so because I don't think your
API is stable yet (I might be wrong). That would let you work "freely"
in the 0.Y.Z version range.

Before accepting/rejecting this PR, we should discuss on how to solve
the current problem and move forward.